Make element type generic
The algorithm works on a pair of lists of any data type.
For example, the data type (element) could be a word (rather than just a
character).
Tokenizing the two given strings and then applying the algorithm on
the two lists of tokens we can compute the word-level distance between
two strings.